✨ Add overlap logic for stashed windows
